[v4] graduate PanelStack2 -> PanelStack component
