+++ packages/core/src/common/classes.ts
@@ -222,11 +222,6 @@ export const PANEL_STACK_HEADER = `${PANEL_STACK}-header`;
 export const PANEL_STACK_HEADER_BACK = `${PANEL_STACK}-header-back`;
 export const PANEL_STACK_VIEW = `${PANEL_STACK}-view`;
 
-export const PANEL_STACK2 = `${NS}-panel-stack2`;
-export const PANEL_STACK2_HEADER = `${PANEL_STACK}-header`;
-export const PANEL_STACK2_HEADER_BACK = `${PANEL_STACK}-header-back`;
-export const PANEL_STACK2_VIEW = `${PANEL_STACK}-view`;
-
 export const POPOVER = `${NS}-popover`;
 export const POPOVER_ARROW = `${POPOVER}-arrow`;
 export const POPOVER_BACKDROP = `${POPOVER}-backdrop`;
