+++ packages/core/src/components/_index.scss
@@ -25,7 +25,6 @@
 @import "overflow-list/overflow-list";
 @import "overlay/overlay";
 @import "panel-stack/panel-stack";
-@import "panel-stack2/panel-stack2";
 @import "popover/popover";
 @import "portal/portal";
 @import "progress-bar/progress-bar";
