+++ packages/core/src/components/components.md
@@ -19,7 +19,6 @@
 @page non-ideal-state
 @page overflow-list
 @page panel-stack
-@page panel-stack2
 @page progress-bar
 @page resize-sensor
 @page skeleton
