+++ packages/core/src/components/panel-stack/_panel-stack.scss
@@ -4,12 +4,12 @@
 @import "../../common/variables";
 @import "~@blueprintjs/core/src/common/react-transition";
 
-.#{$ns}-panel-stack2 {
+.#{$ns}-panel-stack {
   overflow: hidden;
   position: relative;
 }
 
-.#{$ns}-panel-stack2-header {
+.#{$ns}-panel-stack-header {
   align-items: center;
   box-shadow: 0 1px $pt-divider-black;
   display: flex;
@@ -33,7 +33,7 @@
   }
 }
 
-.#{$ns}-button.#{$ns}-panel-stack2-header-back {
+.#{$ns}-button.#{$ns}-panel-stack-header-back {
   margin-left: $pt-grid-size / 2;
   padding-left: 0;
   white-space: nowrap;
@@ -44,7 +44,7 @@
   }
 }
 
-.#{$ns}-panel-stack2-view {
+.#{$ns}-panel-stack-view {
   @include position-all(absolute, 0);
 
   background-color: $white;
@@ -67,16 +67,16 @@
 }
 
 // PUSH transition: enter from right (100%), existing panel moves off left.
-.#{$ns}-panel-stack2-push {
+.#{$ns}-panel-stack-push {
   @include react-transition-phase(
-    "#{$ns}-panel-stack2",
+    "#{$ns}-panel-stack",
     "enter",
     (transform: translateX(100%) translate(0%), opacity: 0 1),
     $easing: ease,
     $duration: $pt-transition-duration * 4
   );
   @include react-transition-phase(
-    "#{$ns}-panel-stack2",
+    "#{$ns}-panel-stack",
     "exit",
     (transform: translateX(-50%) translate(0%), opacity: 0 1),
     $easing: ease,
@@ -85,16 +85,16 @@
 }
 
 // POP transition: enter from left (-50%), existing panel moves off right.
-.#{$ns}-panel-stack2-pop {
+.#{$ns}-panel-stack-pop {
   @include react-transition-phase(
-    "#{$ns}-panel-stack2",
+    "#{$ns}-panel-stack",
     "enter",
     (transform: translateX(-50%) translate(0%), opacity: 0 1),
     $easing: ease,
     $duration: $pt-transition-duration * 4
   );
   @include react-transition-phase(
-    "#{$ns}-panel-stack2",
+    "#{$ns}-panel-stack",
     "exit",
     (transform: translateX(100%) translate(0%), opacity: 0 1),
     $easing: ease,
