+++ packages/core/src/components/panel-stack/panelStack.tsx
@@ -20,13 +20,13 @@ import { CSSTransition, TransitionGroup } from "react-transition-group";
 
 import { Classes, DISPLAYNAME_PREFIX, Props } from "../../common";
 import { Panel } from "./panelTypes";
-import { PanelView2 } from "./panelView2";
+import { PanelView } from "./panelView";
 
 /**
  * @template T type union of all possible panels in this stack
  */
 // eslint-disable-next-line @typescript-eslint/ban-types
-export interface PanelStack2Props<T extends Panel<object>> extends Props {
+export interface PanelStackProps<T extends Panel<object>> extends Props {
     /**
      * The initial panel to show on mount. This panel cannot be removed from the
      * stack and will appear when the stack is empty.
@@ -70,12 +70,12 @@ export interface PanelStack2Props<T extends Panel<object>> extends Props {
     stack?: T[];
 }
 
-interface PanelStack2Component {
+interface PanelStackComponent {
     /**
      * @template T type union of all possible panels in this stack
      */
     // eslint-disable-next-line @typescript-eslint/ban-types
-    <T extends Panel<object>>(props: PanelStack2Props<T>): JSX.Element | null;
+    <T extends Panel<object>>(props: PanelStackProps<T>): JSX.Element | null;
     displayName: string;
 }
 
@@ -83,7 +83,7 @@ interface PanelStack2Component {
  * @template T type union of all possible panels in this stack
  */
 // eslint-disable-next-line @typescript-eslint/ban-types
-export const PanelStack2: PanelStack2Component = <T extends Panel<object>>(props: PanelStack2Props<T>) => {
+export const PanelStack: PanelStackComponent = <T extends Panel<object>>(props: PanelStackProps<T>) => {
     const { renderActivePanelOnly = true, showPanelHeader = true } = props;
     const [direction, setDirection] = React.useState("push");
 
@@ -134,8 +134,8 @@ export const PanelStack2: PanelStack2Component = <T extends Panel<object>>(props
             const key = renderActivePanelOnly ? stack.length : layer;
 
             return (
-                <CSSTransition classNames={Classes.PANEL_STACK2} key={key} timeout={400}>
-                    <PanelView2<T>
+                <CSSTransition classNames={Classes.PANEL_STACK} key={key} timeout={400}>
+                    <PanelView<T>
                         onClose={handlePanelClose}
                         onOpen={handlePanelOpen}
                         panel={panel}
@@ -147,7 +147,7 @@ export const PanelStack2: PanelStack2Component = <T extends Panel<object>>(props
         })
         .reverse();
 
-    const classes = classNames(Classes.PANEL_STACK2, `${Classes.PANEL_STACK2}-${direction}`, props.className);
+    const classes = classNames(Classes.PANEL_STACK, `${Classes.PANEL_STACK}-${direction}`, props.className);
 
     return (
         <TransitionGroup className={classes} component="div">
@@ -155,4 +155,4 @@ export const PanelStack2: PanelStack2Component = <T extends Panel<object>>(props
         </TransitionGroup>
     );
 };
-PanelStack2.displayName = `${DISPLAYNAME_PREFIX}.PanelStack2`;
+PanelStack.displayName = `${DISPLAYNAME_PREFIX}.PanelStack`;
