--- packages/core/src/components/panel-stack/tmp.html
@@ -1,1 +1,0 @@
-<div class="bp3-panel-stack2 bp3-panel-stack2-push"><div class="bp3-panel-stack-view bp3-panel-stack2-enter bp3-panel-stack2-enter-active"><div class="bp3-panel-stack-header"><span><button type="button" class="bp3-button bp3-minimal bp3-small bp3-panel-stack-header-back"><span icon="chevron-left" class="bp3-icon bp3-icon-chevron-left"><svg data-icon="chevron-left" width="16" height="16" viewBox="0 0 16 16"><desc>chevron-left</desc><path d="M7.41 8l3.29-3.29c.19-.18.3-.43.3-.71a1.003 1.003 0 00-1.71-.71l-4 4C5.11 7.47 5 7.72 5 8c0 .28.11.53.29.71l4 4a1.003 1.003 0 001.42-1.42L7.41 8z" fill-rule="evenodd"></path></svg></span><span class="bp3-button-text">New Panel 1</span></button></span><div class="bp3-text-overflow-ellipsis bp3-heading">Test Title</div><span></span></div><div><button id="new-panel-button"></button><button id="close-panel-button"></button></div></div><div class="bp3-panel-stack-view bp3-panel-stack2-exit bp3-panel-stack2-exit-active"><div class="bp3-panel-stack-header"><span></span><div class="bp3-text-overflow-ellipsis bp3-heading">Test Title</div><span></span></div><div><button id="new-panel-button"></button><button id="close-panel-button"></button></div></div></div>
