+++ packages/docs-app/src/examples/core-examples/index.ts
@@ -47,8 +47,7 @@ export * from "./numericInputExtendedExample";
 export * from "./nonIdealStateExample";
 export * from "./overflowListExample";
 export * from "./overlayExample";
-export * from "./panelStackExample";
-export { PanelStack2Example } from "./panelStack2Example";
+export { PanelStackExample } from "./panelStackExample";
 export * from "./popoverDismissExample";
 export * from "./popoverExample";
 export * from "./popoverInteractionKindExample";
