+++ packages/docs-app/src/styles/_examples.scss
@@ -307,7 +307,6 @@
   }
 }
 
-#{example("PanelStack2Example")},
 #{example("PanelStackExample")} {
   .docs-panel-stack-example {
     box-shadow: $pt-elevation-shadow-0;
